\name{cac40}
\docType{data}
\alias{cac40}
\title{
  CAC 40 Stock Market Index (France) as timeSeries object from January 1994 to March 25, 2004
}
\description{
This timeSeries data set provides the daily closing values of the French CAC 40 
stock index for the period 1994 to March 2004 in its cac40@Data slot.
QRMlib's R-version 1.4.2 and above 
supplies data in both timeSeries and data.frame versions.
}
\usage{
data(cac40)
}
\format{
This timeSeries object contains the prices for the index at cac40@Data and the corresponding dates 
at cac40@positions.  
}
\seealso{
   \code{\link{cac40.df}}
}
\keyword{datasets}


