\name{danish}
\alias{danish}
\title{
  Danish Data from January 1980 through December 1990 as timeSeries object 
}
\description{
The \code{danish} timeSeries dataset provides the daily closing value for the Danish fire losses 
measured daily from January 1980 through December 1990 in its danish@Data slot. QRMlib's R-version 
1.4.2 and above supply data in both timeSeries and data.frame versions.
}
\usage{
data(danish)
}
\format{
This timeSeries object contains the Fire losses (in million of kroner) at danish@Data and the 
corresponding dates at danish@positions.  
}
\seealso{
   \code{\link{danish.df}}
}
\keyword{datasets}


