\name{dji}
\alias{dji}
\docType{data}
\title{
  Dow Jones Index (timeSeries Object) January 2, 1980-March 25, 2004
}
\description{
The \code{dji} timeSeries dataset provides the daily closing value for the Dow Jones index 
from January 1980 to March 2004 in its dji@Data slot. QRMlib's R-version 1.4.2 and above 
supplies data in both timeSeries and data.frame versions.
}
\usage{
data(dji)
}
\format{
This timeSeries object contains the prices for the index at dji@Data and the corresponding dates 
at dji@positions.  
}
\seealso{
   \code{\link{dji.df}}
}
\keyword{datasets}
