\name{ElogGIG}
\alias{ElogGIG}
\title{
  Log Moment of GIG 
}
\description{
calculates log moment of generalized hyperbolic distribution 
}
\usage{
ElogGIG(lambda, chi, psi)
}
\arguments{
\item{lambda}{
lambda parameter 
}
\item{chi}{
chi parameter 
}
\item{psi}{
psi parameter 
}
}
\value{
log moment 
}
\details{
see page 497 of QRM 
}
\seealso{
   \code{\link{rGIG}},
   \code{\link{EGIG}}
}

\keyword{distribution}

