\name{ftse100}
\alias{ftse100}
\docType{data}
\title{
  FTSE 100 Stock Market Index as timeSeries object
}
\description{
The \code{ftse100} timeSeries dataset provides the daily closing value for the FTSE index 
from January 1980 to March 2004 in its ftse100@Data slot. QRMlib's R-version 1.4.2 and above 
supplies data in both timeSeries and data.frame versions.
}
\usage{
data(ftse100)
}
\format{
This timeSeries object contains the prices for the index at ftse100@Data and the corresponding dates 
at ftse100@positions.  
}
\seealso{
   \code{\link{ftse100.df}} 
}
\keyword{datasets}
