\name{FXGBP.RAW}
\alias{FXGBP.RAW}
\docType{data}
\title{
  Sterling Exchange Rates as timeSeries object 
}
\description{
The \code{FXGBP.RAW} timeSeries dataset provides daily exchange rates 
for major currencies (dollar, yen, euro, Swiss franc) against 
the British pound for the period January 1987 through March 2004 
}
\usage{
data(FXGBP.RAW)
}
\format{
This timeSeries contains the following 4 columns: \cr
\tabular{ll}{
\emph{GBP.USD} \tab exchange rate against U.S. dollar \cr
\emph{GBP.EUR} \tab exchange rate against Euro \cr
\emph{GBP.JYN} \tab exchange rate against Japanese yen   \cr
\emph{GBP.CHF} \tab exchange rate against Swiss Frank \cr
}
There are 4360 rows with daily rates from 1987 to 2004
}
\seealso{
   \code{\link{FXGBP.RAW.df}}
}
\keyword{datasets}
