\name{hsi}
\alias{hsi}
\docType{data}
\title{
  Hang Seng Stock Market Index (timeSeries) 
}
\description{
The \code{hsi} timeSeries dataset provides the daily closing value for the Hanh Seng Index 
from January 1994 to March 2004 in its hsi@Data slot. QRMlib's R-version 1.4.2 and above 
supplies data in both timeSeries and data.frame versions.
}
\usage{
data(hsi)
}
\format{
This timeSeries object contains the prices for the index at hsi@Data and the corresponding dates 
at hsi@positions.  
}
\seealso{
   \code{\link{hsi.df}}
This timeSeries data set provides the daily closing values 
}
\keyword{datasets}

