\name{lbeta}
\alias{lbeta}
\title{
  Log Beta Function 
}
\description{
calculates logarithm of beta function 
}
\usage{
lbeta(a, b)
}
\arguments{
\item{a}{
vector of values of argument 1 
}
\item{b}{
vector of values of argument 2 
}
}
\value{
vector of values of logarithm of beta function 
}
\seealso{
  For other special functions, see \code{\link{Special}}
}
\keyword{math}

