\name{nasdaq}
\docType{data}
\alias{nasdaq}
\title{
  NASDAQ Stock Market Index (timeSeries object) January 3, 1994 to March 25, 2004
} 
\description{
The \code{nasdaq} timeSeries dataset provides the daily closing value for the daily 
closing values of the NASDAQ index from January 1994 to March 2004 in its nasdaq@Data slot. 
QRMlib's R-version 1.4.2 and above  supplies data in both timeSeries and data.frame versions.
}
\usage{
data(nasdaq)
}
\format{
This timeSeries object contains the prices for the index at nasdaq@Data and the corresponding dates 
at nasdaq@positions.  
}
\seealso{
   \code{\link{nasdaq.df}}
}
\keyword{datasets}

