\name{nikkei}
\docType{data}
\alias{nikkei}
\title{
  Nikkei Stock Market Index (timeSeries Object) January 4, 1994-March 25, 2004
}
\description{
The \code{nikkei} timeSeries dataset provides the daily closing value for the Nikkei index 
from January 1994 to March 2004 in its nikkei@Data slot. QRMlib's R-version 1.4.2 and above 
supplies data in both timeSeries and data.frame versions.
}
\usage{
data(nikkei)
}
\format{
This timeSeries object contains the prices for the index at nikkei@Data and the corresponding dates 
at nikkei@positions.  
}
\seealso{
   \code{\link{nikkei.df}}
}
\keyword{datasets}

