\name{psifunc}
\alias{psifunc}
\title{
  Psi or Digamma Function 
}
\description{
calculates psi or digamma function 
}
\usage{
psifunc(x=2, logvalue=FALSE)
}
\arguments{
\item{x}{
vector of values at which to calculate function 
}
\item{logvalue}{
whether logarithm of function should be returned; default is FALSE 
}
}
\value{
vector of values of psi or digamma function 
}
\seealso{
   \code{\link{besselM3}}
}
\keyword{math}
