\name{smi}
\docType{data}
\alias{smi}
\title{
  Swiss Market Index (timeSeries Object) November 9, 1990 to March 25, 2004
}
\description{
The \code{smi} timeSeries dataset provides the daily closing value for the Swiss Market index 
from November 1990 to March 2004 in its smii@Data slot. QRMlib's R-version 1.4.2 and above 
supplies data in both timeSeries and data.frame versions.
}
\usage{
data(smi)
}
\format{
This timeSeries object contains the prices for the index at smi@Data and the corresponding dates 
at smi@positions.  
}
\seealso{
   \code{\link{smi.df}}
}
\keyword{datasets}

