\name{sp500}
\docType{data}
\alias{sp500}
\title{
  Standard and Poors 500 Index (timeSeries Object) January 2, 1990-March 25, 2004
}
\description{
The \code{sp500} timeSeries dataset provides the daily closing value for the S and P 500 Index 
from January 1980 to March 2004 in its sp500@Data slot. QRMlib's R-version 1.4.2 and above 
supplies data in both timeSeries and data.frame versions.
}
\usage{
data(dji)
}
\format{
This timeSeries object contains the prices for the index at sp500@Data and the corresponding dates 
at sp500@positions.  
}
\seealso{
   \code{\link{sp500.df}}
}
\keyword{datasets}
