\name{xdax}
\docType{data}
\alias{xdax}
\title{
  Xetra DAX German Index (timeSeries Object) January 3, 1994-March 25, 2004
}
\description{
The \code{xdax} timeSeries dataset provides the daily closing value for the German Xextra DAX index 
from January 1994 to March 2004 in its xdax@Data slot. QRMlib's R-version 1.4.2 and above 
supplies data in both timeSeries and data.frame versions.
}
\usage{
data(xdax)
}
\format{
This timeSeries object contains the prices for the index at xdax@Data and the corresponding dates 
at xdax@positions.  
}
\seealso{
   \code{\link{xdax.df}}
}
\keyword{datasets}

